+++ co/index.d.ts
@@ -1,1 +1,0 @@
-export function Async(target: any, name:string);