+++ mongoose/src/index.ts
@@ -1,2 +1,3 @@
 export * from './decorators';
-export { bootstrapMongoose } from './mongoose';
+export { bootstrapMongoose, ref } from './mongoose';
+export { ModelClass } from './model-class';