+++ playground/co/index.ts
@@ -1,4 +1,4 @@
-import { Async } from '../';
+import { Async } from '@decorators/co';
 
 let testAsyncFunc = () => {
   return new Promise(resolve => {
@@ -11,11 +11,12 @@ let testAsyncFunc = () => {
 
 class TestController {
 
-  @Async
+  @Async()
   *getData() {
     console.log('code before async function');
     yield testAsyncFunc();
     console.log('code after async function');
+    process.exit(0);
   }
 
 }